+++ openapi/src/api/types.rs
@@ -1,0 +1,5 @@
+pub mod job;
+pub mod workflow;
+
+pub use job::*;
+pub use workflow::*;
